+++ src/main/java/service/local/create/interface_adapter/CreateLocalVaultController.java
@@ -1,4 +1,4 @@
-package service.local.interface_adapter;
+package service.local.create.interface_adapter;
 
 import javax.swing.JFileChooser;
